remove the unnecessary `src/enums.ts`
