+++ src/compile/stack.ts
@@ -3,7 +3,8 @@ import {Config} from '../config';
 import {FieldDef} from '../fielddef';
 import {Model, ScaleMap} from './Model';
 import {Channel, X, Y, COLOR, DETAIL, ORDER} from '../channel';
-import {ScaleType, StackOffset} from '../enums';
+import {ScaleType} from '../scale';
+import {StackOffset} from '../config';
 import {BAR, AREA, Mark} from '../mark';
 import {field, isMeasure} from '../fielddef';
 import {has, isAggregate} from '../encoding';
